+++ clang/docs/ReleaseNotes.rst
@@ -463,6 +463,7 @@ OpenMP Support
   modifier in the ``adjust_args`` clause.
 - Allow array length to be omitted in array section subscript expression.
 - Fixed non-contiguous strided update in the ``omp target update`` directive with the ``from`` clause.
+- Added support for threadset clause in task and taskloop directives.
 
 Improvements
 ^^^^^^^^^^^^